Initialize SampleError class
